proj/requirements.txt
Django>=1.8,<2.0
djangorestframework
psycopg2
sphinxapi-py3
djangorestframework-jwt
django-sphinxql
